pyproject.toml
[tool.poetry]
name = "tweetscience"
version = "1.0.0"
description = ""
authors = ["Jon Besga <jonan.bsg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.4.2"
python-twitter = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
